Polimorfismo

+++ src/Interface/ImpactoAmbiental.java
@@ -1,5 +1,5 @@
 package Interface;
 
 public interface ImpactoAmbiental {
-    double calcularImpacto();
+    int calcularImpacto();
 }
